+++ contrib/bitcoin-qt.pro
@@ -4,6 +4,7 @@ FORMS += \
     ../src/qt/forms/askpassphrasedialog.ui \
     ../src/qt/forms/coincontroldialog.ui \
     ../src/qt/forms/editaddressdialog.ui \
+    ../src/qt/forms/helpmessagedialog.ui \
     ../src/qt/forms/intro.ui \
     ../src/qt/forms/openuridialog.ui \
     ../src/qt/forms/optionsdialog.ui \
