--- src/qt/aboutdialog.h
@@ -1,34 +1,0 @@
-// Copyright (c) 2011-2013 The Bitcoin developers
-// Distributed under the MIT/X11 software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
-
-#ifndef ABOUTDIALOG_H
-#define ABOUTDIALOG_H
-
-#include <QDialog>
-
-class ClientModel;
-
-namespace Ui {
-    class AboutDialog;
-}
-
-/** "About" dialog box */
-class AboutDialog : public QDialog
-{
-    Q_OBJECT
-
-public:
-    explicit AboutDialog(QWidget *parent);
-    ~AboutDialog();
-
-    void setModel(ClientModel *model);
-
-private:
-    Ui::AboutDialog *ui;
-
-private slots:
-    void on_buttonBox_accepted();
-};
-
-#endif // ABOUTDIALOG_H
