+++ src/qt/guiutil.h
@@ -123,26 +123,6 @@ namespace GUIUtil
     /** Restore window size and position */
     void restoreWindowGeometry(const QString& strSetting, const QSize &defaultSizeIn, QWidget *parent);
 
-    /** Help message for Bitcoin-Qt, shown with --help. */
-    class HelpMessageBox : public QMessageBox
-    {
-        Q_OBJECT
-
-    public:
-        HelpMessageBox(QWidget *parent = 0);
-
-        /** Show message box or print help message to standard output, based on operating system. */
-        void showOrPrint();
-
-        /** Print help message to console */
-        void printToConsole();
-
-    private:
-        QString header;
-        QString coreOptions;
-        QString uiOptions;
-    };
-
 } // namespace GUIUtil
 
 #endif // GUIUTIL_H
