+++ src/qt/rpcconsole.cpp
@@ -7,6 +7,7 @@
 
 #include "clientmodel.h"
 #include "guiutil.h"
+#include "utilitydialog.h"
 
 #include "rpcserver.h"
 #include "rpcclient.h"
@@ -443,8 +444,8 @@ void RPCConsole::scrollToEnd()
 
 void RPCConsole::on_showCLOptionsButton_clicked()
 {
-    GUIUtil::HelpMessageBox help;
-    help.exec();
+    HelpMessageDialog *help = new HelpMessageDialog(this);
+    help->show();
 }
 
 void RPCConsole::on_sldGraphRange_valueChanged(int value)
